software/train/requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.24
Pillow>=9.0